test(gui): the offer's own defect is guarded, not just described
guard-prover, round two. The three mutations that made every
Desktop row vanish while the suite stayed green are red now. It
found three more, and two were the fixes themselves being
unguarded.

The container swap this branch exists to remove was fully
re-introducible: nest `if bound { … } else { … }` INSIDE the
`hasRows` guard and every positive needle still matches — one
`SettingsDisclosure(`, the mount spelled verbatim, the `onAppear`
present. A substring set cannot see structure. `body` is now read
as a brace-balanced run and held to one branch drawing `families`
once, which reds on the swap twice over.

The bridge filter was watched by nothing at all — deleting it, or
emptying the register it reads, passed the full suite. That was
one of the three defects this branch fixes, silently revertible.
It has a behavioural test over the built index now, both ways.

And this branch had re-opened the dead-declaration guard it just
closed: the rename moved the collision off `ShortcutsRowOrder`,
but the search index and its synonym table then mentioned the
same ids, so a REGISTRATION stood in for a RENDER and pointing a
mount at the wrong drawer stayed green. Those two files are
registries rather than renderers and are named as such.
